@@ -34,20 +34,20 @@
     "checksum": "^0.1.1",
     "chron-time": "^0.0.4",
     "columnify": "^1.5.4",
-    "compression-webpack-plugin": "^1.1.11",
-    "core-js": "2",
-    "css-loader": "^1.0.0",
+    "compression-webpack-plugin": "^2.0.0",
+    "core-js": "^2.6.5",
+    "css-loader": "^2.1.1",
     "debounce": "^1.0.2",
-    "deepmerge": "^2.1.0",
+    "deepmerge": "^3.2.0",
     "emailjs": "^2.1.0",
     "expand-var": "^0.1.3",
     "extract-text-webpack-plugin": "4.0.0-beta.0",
-    "file-loader": "^2.0.0",
+    "file-loader": "^3.0.1",
     "file-uploader": "^1.0.0",
     "fs-extra": "^7.0.0",
     "fs-extra-promise": "^1.0.1",
     "glob-promise": "^3.4.0",
-    "immutability-helper": "^2.6.6",
+    "immutability-helper": "^3.0.0",
     "jsx-control-statements": "^3.1.5",
     "koa": "^2.7.0",
     "koa-bodyparser": "^4.1.0",
@@ -70,15 +70,15 @@
     "postcss-import": "^12.0.0",
     "postcss-loader": "^3.0.0",
     "postcss-mixins": "^6.2.0",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-dropzone": "^5.0.1",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-dropzone": "^10.1.4",
     "react-leaflet": "^2.2.1",
     "react-notification": "^6.8.2",
     "react-region-select": "^2.2.0",
-    "react-router-dom": "^4.3.1",
+    "react-router-dom": "^5.0.0",
     "react-semantic-ui-range": "^0.6.1",
-    "react-svg-loader": "^2.1.0",
+    "react-svg-loader": "^3.0.0",
     "request": "^2.85.0",
     "request-promise-native": "^1.0.7",
     "semantic-ui-css": "^2.3.1",
@@ -90,10 +90,10 @@
     "simple-xml2json": "^1.2.3",
     "source-map-support": "^0.5.12",
     "store": "^2.0.12",
-    "strip-ansi": "^4.0.0",
+    "strip-ansi": "^5.2.0",
     "style-loader": "^0.23.0",
-    "superagent": "^3.8.2",
-    "uglify-js": "^3.5.9",
+    "superagent": "^5.0.5",
+    "uglify-js": "^3.5.10",
     "url-loader": "^1.0.1",
     "uuid": "^3.2.1",
     "walk-promise": "^0.2.0",
@@ -107,7 +107,7 @@
     "eslint": "^5.16.0",
     "eslint-plugin-babel": "^5.0.0",
     "eslint-plugin-jsx-control-statements": "^2.2.0",
-    "eslint-plugin-react": "^7.12.4",
+    "eslint-plugin-react": "^7.13.0",
     "mocha": "^6.1.4"
   },
   "engines": {
