@@ -1,8 +1,9 @@
 
 import React from "react";
 import PropTypes from "prop-types";
-import Component from "lib/component";
+import { withRouter } from "react-router-dom";
 import { Header } from "semantic-ui-react";
+import Component from "lib/component";
 import SidebarMenuItem from "./SidebarMenuItem";
 
 class Sidebar extends Component {
@@ -15,31 +16,26 @@ class Sidebar extends Component {
                         theme={this.props.theme}
                         text="News"
                         path={`${this.props.match.path}/news`}
-                        location={this.props.location}
                     />
                     <SidebarMenuItem
                         theme={this.props.theme}
                         text="Search"
                         path={`${this.props.match.path}/search`}
-                        location={this.props.location}
                     />
                     <SidebarMenuItem
                         theme={this.props.theme}
                         text="Browse by name"
                         path={`${this.props.match.path}/name`}
-                        location={this.props.location}
                     />
                     <SidebarMenuItem
                         theme={this.props.theme}
                         text="Browse by label"
                         path={`${this.props.match.path}/label`}
-                        location={this.props.location}
                     />
                     <SidebarMenuItem
                         theme={this.props.theme}
                         text="Browse albums by year"
                         path={`${this.props.match.path}/year`}
-                        location={this.props.location}
                     />
                 </div>
                 <div className={this.props.theme.menu}>
@@ -48,7 +44,6 @@ class Sidebar extends Component {
                         theme={this.props.theme}
                         text="Charts"
                         path={`${this.props.match.path}/chart`}
-                        location={this.props.location}
                     />
                 </div>
             </div>
@@ -58,8 +53,7 @@ class Sidebar extends Component {
 
 Sidebar.propTypes = {
     theme: PropTypes.object.isRequired,
-    match: PropTypes.object.isRequired,
-    location: PropTypes.object.isRequired
+    match: PropTypes.object.isRequired
 };
 
-export default Sidebar;
+export default withRouter(Sidebar);
