@@ -229,8 +229,7 @@ class About extends Component {
 
 About.propTypes = {
     theme: PropTypes.object,
-    node: PropTypes.object.isRequired,
-    match: PropTypes.object.isRequired
+    node: PropTypes.object.isRequired
 };
 
 export default About;
