@@ -53,12 +53,7 @@ class Organize extends Component {
 
 Organize.propTypes = {
     theme: PropTypes.object,
-    node: PropTypes.object.isRequired,
-    match: PropTypes.object.isRequired
-};
-
-Organize.contextTypes = {
-    router: PropTypes.object.isRequired
+    node: PropTypes.object.isRequired
 };
 
 export default Organize;
