@@ -2,17 +2,14 @@
 import React from "react";
 import PropTypes from "prop-types";
 import Component from "lib/component";
+import { Link } from "react-router-dom";
 import NodeIcon from "./NodeIcon";
 
 class NodeLink extends Component {
-    onClick = () => {
-        this.context.router.history.push(`/node${this.props.node.path}`);
-    }
-
     render() {
         return (
-            <a
-                onClick={this.onClick}
+            <Link
+                to={`/node${this.props.node.path}`}
                 className={this.classNames(this.props.theme.nodeLink, this.props.className)}
             >
                 <If condition={this.props.icon}>
@@ -22,7 +19,7 @@ class NodeLink extends Component {
                     />
                 </If>
                 {this.props.node.attributes.name}
-            </a>
+            </Link>
         );
     }
 }
@@ -32,10 +29,6 @@ NodeLink.propTypes = {
     className: PropTypes.string,
     node: PropTypes.object.isRequired,
     icon: PropTypes.bool
-};
-
-NodeLink.contextTypes = {
-    router: PropTypes.object.isRequired
 };
 
 export default NodeLink;
